split place search into address search and named search
The presence/absence of houenumbers makes quite a difference for search.
