+++ src/nominatim_api/search/db_searches/place_search.py
@@ -5,14 +5,14 @@
 # Copyright (C) 2025 by the Nominatim developer community.
 # For a full list of authors see the git log.
 """
-Implementation of search for a named place.
+Implementation of search for a named place (without housenumber).
 """
-from typing import cast, List, AsyncIterator
+from typing import cast
 
 import sqlalchemy as sa
 
 from . import base
-from ...typing import SaBind, SaExpression, SaColumn, SaFromClause, SaScalarSelect
+from ...typing import SaBind, SaExpression, SaColumn
 from ...types import SearchDetails, Bbox
 from ...sql.sqlalchemy_types import Geometry
 from ...connection import SearchConnection
@@ -30,117 +30,16 @@ NEAR_RADIUS_PARAM: SaBind = sa.bindparam('near_radius')
 COUNTRIES_PARAM: SaBind = sa.bindparam('countries')
 
 
-def _int_list_to_subquery(inp: List[int]) -> 'sa.Subquery':
-    """ Create a subselect that returns the given list of integers
-        as rows in the column 'nr'.
-    """
-    vtab = sa.func.JsonArrayEach(sa.type_coerce(inp, sa.JSON))\
-             .table_valued(sa.column('value', type_=sa.JSON))
-    return sa.select(sa.cast(sa.cast(vtab.c.value, sa.Text), sa.Integer).label('nr')).subquery()
-
-
-def _interpolated_position(table: SaFromClause, nr: SaColumn) -> SaColumn:
-    pos = sa.cast(nr - table.c.startnumber, sa.Float) / (table.c.endnumber - table.c.startnumber)
-    return sa.case(
-            (table.c.endnumber == table.c.startnumber, table.c.linegeo.ST_Centroid()),
-            else_=table.c.linegeo.ST_LineInterpolatePoint(pos)).label('centroid')
-
-
-def _make_interpolation_subquery(table: SaFromClause, inner: SaFromClause,
-                                 numerals: List[int], details: SearchDetails) -> SaScalarSelect:
-    all_ids = sa.func.ArrayAgg(table.c.place_id)
-    sql = sa.select(all_ids).where(table.c.parent_place_id == inner.c.place_id)
-
-    if len(numerals) == 1:
-        sql = sql.where(sa.between(numerals[0], table.c.startnumber, table.c.endnumber))\
-                 .where((numerals[0] - table.c.startnumber) % table.c.step == 0)
-    else:
-        sql = sql.where(sa.or_(
-                *(sa.and_(sa.between(n, table.c.startnumber, table.c.endnumber),
-                          (n - table.c.startnumber) % table.c.step == 0)
-                  for n in numerals)))
-
-    if details.excluded:
-        sql = sql.where(base.exclude_places(table))
-
-    return sql.scalar_subquery()
-
-
-async def _get_placex_housenumbers(conn: SearchConnection,
-                                   place_ids: List[int],
-                                   details: SearchDetails) -> AsyncIterator[nres.SearchResult]:
-    t = conn.t.placex
-    sql = base.select_placex(t).add_columns(t.c.importance)\
-                               .where(t.c.place_id.in_(place_ids))
-
-    if details.geometry_output:
-        sql = base.add_geometry_columns(sql, t.c.geometry, details)
-
-    for row in await conn.execute(sql):
-        result = nres.create_from_placex_row(row, nres.SearchResult)
-        assert result
-        result.bbox = Bbox.from_wkb(row.bbox)
-        yield result
-
-
-async def _get_osmline(conn: SearchConnection, place_ids: List[int],
-                       numerals: List[int],
-                       details: SearchDetails) -> AsyncIterator[nres.SearchResult]:
-    t = conn.t.osmline
-
-    values = _int_list_to_subquery(numerals)
-    sql = sa.select(t.c.place_id, t.c.osm_id,
-                    t.c.parent_place_id, t.c.address,
-                    values.c.nr.label('housenumber'),
-                    _interpolated_position(t, values.c.nr),
-                    t.c.postcode, t.c.country_code)\
-            .where(t.c.place_id.in_(place_ids))\
-            .join(values, values.c.nr.between(t.c.startnumber, t.c.endnumber))
-
-    if details.geometry_output:
-        sub = sql.subquery()
-        sql = base.add_geometry_columns(sa.select(sub), sub.c.centroid, details)
-
-    for row in await conn.execute(sql):
-        result = nres.create_from_osmline_row(row, nres.SearchResult)
-        assert result
-        yield result
-
-
-async def _get_tiger(conn: SearchConnection, place_ids: List[int],
-                     numerals: List[int], osm_id: int,
-                     details: SearchDetails) -> AsyncIterator[nres.SearchResult]:
-    t = conn.t.tiger
-    values = _int_list_to_subquery(numerals)
-    sql = sa.select(t.c.place_id, t.c.parent_place_id,
-                    sa.literal('W').label('osm_type'),
-                    sa.literal(osm_id).label('osm_id'),
-                    values.c.nr.label('housenumber'),
-                    _interpolated_position(t, values.c.nr),
-                    t.c.postcode)\
-            .where(t.c.place_id.in_(place_ids))\
-            .join(values, values.c.nr.between(t.c.startnumber, t.c.endnumber))
-
-    if details.geometry_output:
-        sub = sql.subquery()
-        sql = base.add_geometry_columns(sa.select(sub), sub.c.centroid, details)
-
-    for row in await conn.execute(sql):
-        result = nres.create_from_tiger_row(row, nres.SearchResult)
-        assert result
-        yield result
-
-
 class PlaceSearch(base.AbstractSearch):
-    """ Generic search for an address or named place.
+    """ Generic search for a named place.
     """
     SEARCH_PRIO = 1
 
     def __init__(self, extra_penalty: float, sdata: SearchData, expected_count: int) -> None:
+        assert not sdata.housenumbers
         super().__init__(sdata.penalty + extra_penalty)
         self.countries = sdata.countries
         self.postcodes = sdata.postcodes
-        self.housenumbers = sdata.housenumbers
         self.qualifiers = sdata.qualifiers
         self.lookups = sdata.lookups
         self.rankings = sdata.rankings
@@ -187,7 +86,7 @@ class PlaceSearch(base.AbstractSearch):
                 sql = sql.where(t.c.centroid
                                    .intersects(VIEWBOX_PARAM,
                                                use_index=details.viewbox.area < 0.2))
-            elif not self.postcodes and not self.housenumbers and self.expected_count >= 10000:
+            elif not self.postcodes and self.expected_count >= 10000:
                 sql = sql.where(t.c.centroid
                                    .intersects(VIEWBOX2_PARAM,
                                                use_index=details.viewbox.area < 0.5))
@@ -200,17 +99,14 @@ class PlaceSearch(base.AbstractSearch):
                 sql = sql.where(t.c.centroid
                                  .ST_Distance(NEAR_PARAM) < NEAR_RADIUS_PARAM)
 
-        if self.housenumbers:
-            sql = sql.where(t.c.address_rank.between(16, 30))
-        else:
-            if details.excluded:
-                sql = sql.where(base.exclude_places(t))
-            if details.min_rank > 0:
-                sql = sql.where(sa.or_(t.c.address_rank >= MIN_RANK_PARAM,
-                                       t.c.search_rank >= MIN_RANK_PARAM))
-            if details.max_rank < 30:
-                sql = sql.where(sa.or_(t.c.address_rank <= MAX_RANK_PARAM,
-                                       t.c.search_rank <= MAX_RANK_PARAM))
+        if details.excluded:
+            sql = sql.where(base.exclude_places(t))
+        if details.min_rank > 0:
+            sql = sql.where(sa.or_(t.c.address_rank >= MIN_RANK_PARAM,
+                                   t.c.search_rank >= MIN_RANK_PARAM))
+        if details.max_rank < 30:
+            sql = sql.where(sa.or_(t.c.address_rank <= MAX_RANK_PARAM,
+                                   t.c.search_rank <= MAX_RANK_PARAM))
 
         inner = sql.limit(10000).order_by(sa.desc(sa.text('importance'))).subquery()
 
@@ -221,8 +117,7 @@ class PlaceSearch(base.AbstractSearch):
         # If the query is not an address search or has a geographic preference,
         # preselect most important items to restrict the number of places
         # that need to be looked up in placex.
-        if not self.housenumbers\
-           and (details.viewbox is None or details.bounded_viewbox)\
+        if (details.viewbox is None or details.bounded_viewbox)\
            and (details.near is None or details.near_radius is not None)\
            and not self.qualifiers:
             sql = sql.add_columns(sa.func.first_value(inner.c.penalty - inner.c.importance)
@@ -278,60 +173,12 @@ class PlaceSearch(base.AbstractSearch):
         sql = sql.add_columns(penalty.label('accuracy'))\
                  .order_by(sa.text('accuracy'))
 
-        if self.housenumbers:
-            hnr_list = '|'.join(self.housenumbers.values)
-            inner = sql.where(sa.or_(tsearch.c.address_rank < 30,
-                                     sa.func.RegexpWord(hnr_list, t.c.housenumber)))\
-                       .subquery()
-
-            # Housenumbers from placex
-            thnr = conn.t.placex.alias('hnr')
-            pid_list = sa.func.ArrayAgg(thnr.c.place_id)
-            place_sql = sa.select(pid_list)\
-                          .where(thnr.c.parent_place_id == inner.c.place_id)\
-                          .where(sa.func.RegexpWord(hnr_list, thnr.c.housenumber))\
-                          .where(thnr.c.linked_place_id == None)\
-                          .where(thnr.c.indexed_status == 0)
-
-            if details.excluded:
-                place_sql = place_sql.where(thnr.c.place_id.not_in(sa.bindparam('excluded')))
-            if self.qualifiers:
-                place_sql = place_sql.where(self.qualifiers.sql_restrict(thnr))
-
-            numerals = [int(n) for n in self.housenumbers.values
-                        if n.isdigit() and len(n) < 8]
-            interpol_sql: SaColumn
-            tiger_sql: SaColumn
-            if numerals and \
-               (not self.qualifiers or ('place', 'house') in self.qualifiers.values):
-                # Housenumbers from interpolations
-                interpol_sql = _make_interpolation_subquery(conn.t.osmline, inner,
-                                                            numerals, details)
-                # Housenumbers from Tiger
-                tiger_sql = sa.case((inner.c.country_code == 'us',
-                                     _make_interpolation_subquery(conn.t.tiger, inner,
-                                                                  numerals, details)
-                                     ), else_=None)
-            else:
-                interpol_sql = sa.null()
-                tiger_sql = sa.null()
-
-            unsort = sa.select(inner, place_sql.scalar_subquery().label('placex_hnr'),
-                               interpol_sql.label('interpol_hnr'),
-                               tiger_sql.label('tiger_hnr')).subquery('unsort')
-            sql = sa.select(unsort)\
-                    .order_by(sa.case((unsort.c.placex_hnr != None, 1),
-                                      (unsort.c.interpol_hnr != None, 2),
-                                      (unsort.c.tiger_hnr != None, 3),
-                                      else_=4),
-                              unsort.c.accuracy)
-        else:
-            sql = sql.where(t.c.linked_place_id == None)\
-                     .where(t.c.indexed_status == 0)
-            if self.qualifiers:
-                sql = sql.where(self.qualifiers.sql_restrict(t))
-            if details.layers is not None:
-                sql = sql.where(base.filter_by_layer(t, details.layers))
+        sql = sql.where(t.c.linked_place_id == None)\
+                 .where(t.c.indexed_status == 0)
+        if self.qualifiers:
+            sql = sql.where(self.qualifiers.sql_restrict(t))
+        if details.layers is not None:
+            sql = sql.where(base.filter_by_layer(t, details.layers))
 
         sql = sql.limit(LIMIT_PARAM)
 
@@ -353,33 +200,6 @@ class PlaceSearch(base.AbstractSearch):
             assert result
             result.bbox = Bbox.from_wkb(row.bbox)
             result.accuracy = row.accuracy
-            if self.housenumbers and row.rank_address < 30:
-                if row.placex_hnr:
-                    subs = _get_placex_housenumbers(conn, row.placex_hnr, details)
-                elif row.interpol_hnr:
-                    subs = _get_osmline(conn, row.interpol_hnr, numerals, details)
-                elif row.tiger_hnr:
-                    subs = _get_tiger(conn, row.tiger_hnr, numerals, row.osm_id, details)
-                else:
-                    subs = None
-
-                if subs is not None:
-                    async for sub in subs:
-                        assert sub.housenumber
-                        sub.accuracy = result.accuracy
-                        if not any(nr in self.housenumbers.values
-                                   for nr in sub.housenumber.split(';')):
-                            sub.accuracy += 0.6
-                        results.append(sub)
-
-                # Only add the street as a result, if it meets all other
-                # filter conditions.
-                if (not details.excluded or result.place_id not in details.excluded)\
-                   and (not self.qualifiers or result.category in self.qualifiers.values)\
-                   and result.rank_address >= details.min_rank:
-                    result.accuracy += 1.0  # penalty for missing housenumber
-                    results.append(result)
-            else:
-                results.append(result)
+            results.append(result)
 
         return results
